refactored, reduced lines
