+++ src/App.js
@@ -1,37 +1,74 @@
-import React from 'react';
 import './styles/App.css';
-import Card from './components/Card';
-import store from './redux/store';
 
+import React from 'react';
+import CardHeader from './components/CardHeader';
+import CardBody from './components/CardBody';
+import { connect } from 'react-redux';
 import { selectLocation } from './redux/actions';
-import { Provider } from 'react-redux'
+import weatherAPI from "./services/weatherAPI"
 
+class App extends React.Component {
 
-//store
-// let store = createStore(location);
-// store.subscribe(() => console.log(store.getState()));
-// store.dispatch(selectLocation(2));
-// store.dispatch(selectLocation(3));
-
-
-
-
-function App() {
-
-  // if (!store.getState.weather.) {
-  // store.dispatch(selectLocation(1));
-
-  return (
-    <div className="app">
-
-      <main>
-        <Provider store={store}>
-          {/* <显示时间> */}
-          <Card />
-        </Provider>
-      </main>
-    </div>
-  );
+  constructor(props) {
+    super(props);
+  };
+  componentDidMount() {
+    //default index is 0
+    this.props.selectLocation(0);
+  }
+  render() {
+    // console.log(this.state.forecasts);
+    console.log(this.props);
+    return (
+      <div className="app">
+        < div className="card" >
+          < CardHeader currentWeather={this.props.currentWeather} locationIndex={this.props.locationIndex} selectLocation={this.props.selectLocation} locationOptions={this.props.locationOptions} />
+          <CardBody forecasts={this.props.forecasts} />
+        </div >
+      </div>
+    )
+  }
 }
 
-export default App;
+
+//takes in store state and return infomation as props to your component.
+function mapStateToProps(state) {
+  // console.log("maping state")
+  let loadingMessage = "loading..."
+
+  if (!state.weather.loading) {
+
+    return {
+      locationOptions: weatherAPI.getLocaitonList(),
+      locationIndex: state.weather.locationIndex,
+
+      currentWeather: state.weather.currentWeather,
+      forecasts: state.weather.forecasts,
+
+      loading: false,
+      selectLocation: selectLocation
+
+    };
+  }
+
+  return {
+    locationOptions: [],
+    locationIndex: 0,
+
+    currentWeather: {
+      status: loadingMessage,
+      statusIcon: loadingMessage,
+      humidity: loadingMessage,
+      wind: loadingMessage,
+      temperature: loadingMessage,
+    },
+    forecasts: {},
+    loading: true,
+
+
+  }
+
+}
+
+export default connect(mapStateToProps, { selectLocation })(App);
+//connect passes functions in second param, as an array
