+++ src/_App.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import './styles/App.css';
+import Card from './App';
+import store from './redux/store';
+
+
+
+//store
+// let store = createStore(location);
+// store.subscribe(() => console.log(store.getState()));
+// store.dispatch(selectLocation(2));
+// store.dispatch(selectLocation(3));
+
+
+
+
+function App() {
+
+  // if (!store.getState.weather.) {
+  // store.dispatch(selectLocation(1));
+
+  return (
+    <div className="app">
+
+      <main>
+        {/* <显示时间> */}
+        {/* <Card /> */}
+      </main>
+    </div>
+  );
+}
+
+export default App;
