+++ src/components/Forecast.js
@@ -8,7 +8,9 @@ function Forecast({ date, forecast }) {
   const { status, temperature } = forecast.summary;
   return (
     <li className="forecast" >
-      <h1 className="forecast__day">{moment(date).format("ddd, DD MMM")}</h1>
+      <h1 className="forecast__day">{moment(date).format("ddd")}</h1>
+      <p className="forecast__day">({moment(date).format("DD MMM")})</p>
+
       <p className=" temperature forecast__temperature">{temperature.min}/{temperature.max}</p>
 
       {Object.entries(status).sort().map(
