+++ src/index.js
@@ -1,9 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-
+import { Provider } from 'react-redux'
+import store from './redux/store';
 import App from './App.js'
 import './index.css';
 
 ReactDOM.render((
-  <App />
+  <Provider store={store}>
+    <App />
+  </Provider>
+
 ), document.getElementById('root'))
