+++ src/styles/App.css
@@ -198,6 +198,74 @@
   min-width: 18rem;
 }
 
+.forecasts .forecast {
+  border-left: 1px solid azure;
+}
+
+.forecasts .forecast {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  min-width: 5rem;
+  -webkit-box-flex: 1;
+      -ms-flex: 1;
+          flex: 1;
+}
+
+.forecasts .forecast__day {
+  text-transform: capitalize;
+  margin-bottom: 0px;
+}
+
+.forecasts .forecast__period {
+  width: 90%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+@media all and (max-width: 920px) {
+  .forecasts .forecast {
+    -ms-flex-preferred-size: 33%;
+        flex-basis: 33%;
+  }
+  .forecasts .forecast:nth-child(2), .forecasts .forecast:nth-child(1) {
+    -ms-flex-preferred-size: 49%;
+        flex-basis: 49%;
+  }
+}
+
+.forecasts .forecast p {
+  -webkit-margin-before: 0.5rem;
+          margin-block-start: 0.5rem;
+  -webkit-margin-after: 0.5rem;
+          margin-block-end: 0.5rem;
+  font-weight: bold;
+  font-size: 1rem;
+  padding-right: 1rem;
+  padding-left: 1rem;
+  text-align: center;
+}
+
+.forecasts .forecast__weatherIcon {
+  min-width: 4rem;
+  max-width: 5vw;
+}
+
 .temperature:after {
   content: "‎°C";
 }
@@ -215,68 +283,5 @@ ul {
   margin: 0;
   padding: 0;
   list-style-type: none;
-}
-
-.forecast {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  min-width: 5rem;
-  -webkit-box-flex: 1;
-      -ms-flex: 1;
-          flex: 1;
-}
-
-.forecast__day {
-  text-transform: capitalize;
-}
-
-.forecast__period {
-  width: 90%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-@media all and (max-width: 920px) {
-  .forecast {
-    -ms-flex-preferred-size: 33%;
-        flex-basis: 33%;
-  }
-  .forecast:nth-child(2), .forecast:nth-child(1) {
-    -ms-flex-preferred-size: 49%;
-        flex-basis: 49%;
-  }
-}
-
-.forecast p {
-  -webkit-margin-before: 0.5rem;
-          margin-block-start: 0.5rem;
-  -webkit-margin-after: 0.5rem;
-          margin-block-end: 0.5rem;
-  font-weight: bold;
-  font-size: 1rem;
-  padding-right: 1rem;
-  padding-left: 1rem;
-  text-align: center;
-}
-
-.forecast__weatherIcon {
-  min-width: 4rem;
-  max-width: 5vw;
 }
 /*# sourceMappingURL=App.css.map */
