+++ src/styles/App.scss
@@ -160,6 +160,51 @@ $sky-background: url(../images/cityBackground/sydney.jpg) 0 / cover fixed;
   // padding-left: 2rem;
   // padding-right: 2rem;
   min-width: $app-min-width;
+
+  .forecast {
+    border-left: 1px solid azure;
+  }
+  .forecast {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    min-width: 5rem;
+    flex: 1;
+    &__day {
+      text-transform: capitalize;
+      margin-bottom: 0px;
+    }
+
+    &__period {
+      width: 90%;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
+
+    @media all and (max-width: 920px) {
+      flex-basis: 33%;
+
+      &:nth-child(2),
+      &:nth-child(1) {
+        flex-basis: 49%;
+      }
+    }
+
+    p {
+      margin-block-start: 0.5rem;
+      margin-block-end: 0.5rem;
+      font-weight: bold;
+      font-size: 1rem;
+      padding-right: 1rem;
+      padding-left: 1rem;
+      text-align: center;
+    }
+    &__weatherIcon {
+      min-width: 4rem;
+      max-width: 5vw;
+    }
+  }
 }
 
 // .delimeter {
@@ -186,45 +231,4 @@ ul {
   margin: 0;
   padding: 0;
   list-style-type: none;
-}
-
-.forecast {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 5rem;
-  flex: 1;
-  &__day {
-    text-transform: capitalize;
-  }
-
-  &__period {
-    width: 90%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  @media all and (max-width: 920px) {
-    flex-basis: 33%;
-
-    &:nth-child(2),
-    &:nth-child(1) {
-      flex-basis: 49%;
-    }
-  }
-
-  p {
-    margin-block-start: 0.5rem;
-    margin-block-end: 0.5rem;
-    font-weight: bold;
-    font-size: 1rem;
-    padding-right: 1rem;
-    padding-left: 1rem;
-    text-align: center;
-  }
-  &__weatherIcon {
-    min-width: 4rem;
-    max-width: 5vw;
-  }
 }
